% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demultiplex.R
\name{demultiplex}
\alias{demultiplex}
\title{demultiplex barcodes and trim primers}
\usage{
demultiplex(x, data, primer_fd, primer_rv, pathout)
}
\arguments{
\item{x}{the list of barcodes}

\item{data}{the output from Input_preprocessed()}

\item{primer_fd}{the string of your forward primers}

\item{primer_rv}{the string of your reverse primers}

\item{pathout}{the path of your output folder (should be empty)}
}
\value{
files of each barcodes before and after trimming primers
}
\description{
demultiplex barcodes and trim primers
}
