% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demultiplex.R
\name{Input_preprocessed}
\alias{Input_preprocessed}
\title{Data preprocess}
\usage{
Input_preprocessed(filepath, format)
}
\arguments{
\item{filepath}{the path of your sequencing data}

\item{format}{fasta or fastq}
}
\value{
dataframe
}
\description{
Data preprocess
}
